+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * `get_paginated`: fix crash when for some reason Magento returns an empty object
+* `save_product`: deprecate `log_response`
+* Deprecate the `logger` constructor parameter
 * `get_paginated`: fix typing for string `fields`
 
 ## 2.8.0 (2025/10/21)